+++ packages/css/src/components/info/info.template.ts
@@ -1,5 +1,6 @@
 import { Info } from '@dso-toolkit/sources';
 import { html, nothing, TemplateResult } from 'lit-html';
+import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 
 import { buttonTemplate } from '../button/button.template';
 
@@ -10,7 +11,10 @@ export function infoTemplate({ fixed, richContent, onClose }: Info<TemplateResul
         ? buttonTemplate({ label: 'Sluiten', variant: null, onClick: onClose, iconMode: 'only' })
         : nothing
       }
-      ${richContent}
+      ${typeof richContent === 'string'
+        ? html`${unsafeHTML(richContent)}`
+        : richContent
+      }
     </div>
   `;
 }