+++ packages/sources/src/components/form/models/form.model.ts
@@ -1,0 +1,8 @@
+import { FormGroup } from './form-group.model';
+
+export interface Form<TemplateFnReturnType> {
+  legend?: string;
+  mode?: 'vertical' | 'horizontal';
+
+  formGroups: FormGroup<TemplateFnReturnType>[] | TemplateFnReturnType;
+}