+++ packages/sources/src/index.ts
@@ -42,6 +42,8 @@ export * from './components/dropdown-menu/dropdown-menu.models';
 export * from './components/dropdown-menu/dropdown-menu.stories';
 export * from './components/footnotes/footnotes.models';
 export * from './components/footnotes/footnotes.stories';
+export * from './components/form/form.models';
+export * from './components/form/form.stories';
 export * from './components/header/header.models';
 export * from './components/header/header.stories';
 export * from './components/helpcenter-panel/helpcenter-panel.models';